add input field in the home page to log attendance by either member id or phone number

+++ androidApp/src/main/java/com/jsramraj/playmatecompanion/android/database/MemberEntity.kt
@@ -9,8 +9,8 @@ import java.util.Date
 @Entity(tableName = "members")
 data class MemberEntity(
     @PrimaryKey
-    val id: Int,
-    val rowNumber: Int,
+    val id: Long,
+    val rowNumber: Long,
     val firstName: String,
     val lastName: String,
     val email: String,
